django/requirements.txt
Django==2.2
django-debug-toolbar==3.2.2
django-redis==5.0.0
djongo==1.3.6
python-engineio==3.11.1
python-socketio==4.4.0
gunicorn==20.0.4
Pillow
sqlparse==0.2.4
requests